Move export_field to field_config
